+++ R/corpus_tidiers.R
@@ -112,16 +112,16 @@ tidy.Corpus <- function(x, collapse = "\n", ...) {
 #' }
 #'
 #' @name corpus_tidiers
-#' @importFrom quanteda texts docvars
+#'
 #' @export
 tidy.corpus <- function(x, ...) {
-  tbl_df(data.frame(text = quanteda::texts(x), quanteda::docvars(x),
-                    stringsAsFactors = FALSE))
+  dplyr::tbl_df(data.frame(text = quanteda::texts(x),
+                           quanteda::docvars(x),
+                           stringsAsFactors = FALSE))
 }
 
 
 #' @rdname corpus_tidiers
-#' @importFrom quanteda metacorpus
 #' @export
 glance.corpus <- function(x, ...) {
   md <- purrr::compact(quanteda::metacorpus(x))